Fix Python 3.8 + python-gnureadline (again?)
Not sure if this actually broke in Python 3.8, or if it broke at some
point with Ansible, or if it never really worked, or what.  This
method for installation will *probably* be more robust than my prior
method, though.  See comment for more details.

One detail that isn't in that comment: now I can't `pip uninstall
gnureadline`---pip doesn't even know I've installed it, due to
`--install-dir`.