Fix macports module's search for installed ports
This was fun: I couldn't figure out why I wasn't getting (IIRC) curl
installed correctly.  The regexp that was used in the old code was
seeing libcurl in the output and returning True as a result.  Now we
use the power of Python's own regexp to process "port installed"
output properly.