Replace ansible_user with ansible_user_id
Apparently ansible_user may not be defined all the time, and
it... refers to the user running Ansible on the Ansible box?  Whereas
ansible_user_id is a gathered fact, containing the *name* of the user
on the remote end?  The documentation was really unclear about this
IMHO, but ansible_user wasn't working, and ansible_user_id does.